changed structure and cleaned up some stuff
